# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: E-Ticaret API
  description: |
    Staj 2026 E-Ticaret projesi backend sözleşmesi (v1).
    Modüller: Auth, Users, Addresses, Categories, Products, Cart, Orders, Favorites.

    ## Genel Kurallar
    - Tüm başarılı yanıtlar `{ "success": true, "data": ... }` zarfı ile döner;
      mutasyon uçları (POST/PUT/PATCH) ayrıca isteğe bağlı bir `message` alanı içerir.
    - `204 No Content` yanıtları HTTP semantiğine uygun olarak gövde içermez
      (ör. DELETE, logout).
    - Tüm hata yanıtları tek bir şema kullanır: `ErrorResponse` (validasyon
      hatalarında `ValidationErrorResponse`). RFC 9457 kullanılmaz.
    - Kaynak yoksa `404`, kaynak var ama iş kuralı ihlali varsa `422`,
      kaynak var ama başka kullanıcıya aitse `403` döner.
    - JWT gerektiren her endpoint için token hiç gönderilmemişse veya süresi
      dolmuşsa/geçersizse `401` döner (bkz. `components/responses/Unauthorized`).
      Bu, ilgili her endpoint'te ayrıca tekrar tekrar açıklanmaz; genel kuraldır.
    - İstemciden gelen JSON'ın bozuk olması (parse edilemeyen syntax), zorunlu bir
      alanın eksik olması veya beklenen tipte olmaması (ör. `quantity` için string
      gönderilmesi) durumunda `400` döner (bkz. `components/responses/BadRequest`).
      Bu da genel bir kuraldır; body veya query parametresi alan her endpoint için
      geçerlidir ve ayrıca tekrarlanmaz.
    - Tüm tarih/saat alanları (`createdAt` vb.) her zaman **UTC, "Z" sonekli**
...
      istemciye aittir.
...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Giriş yapan kullanıcının profili
      responses:
        '200':
          description: Kullanıcı profili
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Profil güncelle (kısmi güncelleme)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                fullName: { type: string, minLength: 2, maxLength: 100 }
                phone:
                  {
                    type: string,
                    pattern: "^\\+?[0-9]{10,15}$",
                    example: '+905551234567',
                  }
      responses:
        '200':
          description: Güncellenmiş profil
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /users/me/password:
    patch:
      operationId: changePassword
      tags: [users]
      summary: Kullanıcının kendi şifresini değiştirmesi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8, maxLength: 72 }
      responses:
        '204':
          description: "Şifre değiştirildi; tüm oturumlar (refresh token'lar) geçersiz kılınır"
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /users/me/email:
    patch:
      operationId: changeEmail
      tags: [users]
      summary: Kullanıcının kendi email adresini değiştirmesi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [newEmail, currentPassword]
              properties:
                newEmail: { type: string, format: email }
                currentPassword: { type: string }
      responses:
        '200':
          description: Email güncellendi
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ==================== ADDRESSES ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT ile korunan endpoint'ler. Kullanıcı yalnızca kendi verisine erişebilir;
        ADMIN gerektiren uçlar x-roles ile işaretlenmiştir. accessToken ~15 dakika,
        refreshToken ~30 gün geçerlidir (bkz. AuthResponse.expiresIn). accessToken
        her zaman `Authorization: Bearer` header'ı ile gönderilir; refreshToken'ın
        teslim şekli istemci tipine göre değişir (bkz. X-Client-Type header'ı).
...
  responses:
    BadRequest:
      description: İstek gövdesi doğrulama hatası içeriyor
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorResponse' }
          example:
            success: false
            message: 'Validation failed'
            statusCode: 400
            status: 'BAD_REQUEST'
            errors:
              - field: 'email'
                message: 'Email is required'
    Unauthorized:
      description: Token yok, geçersiz veya süresi dolmuş
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            success: false
            message: 'Invalid or expired token'
            statusCode: 401
            status: 'UNAUTHORIZED'
...
    Conflict:
      description: Kaynak zaten mevcut / çakışma
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            success: false
            message: 'Resource already exists'
            statusCode: 409
            status: 'CONFLICT'
...
    InternalServerError:
      description: Beklenmeyen sunucu hatası
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            success: false
            message: 'An unexpected error occurred'
            statusCode: 500
            status: 'INTERNAL_SERVER_ERROR'
  schemas:
...
    ErrorResponse:
      type: object
      description: >
        Tüm hata yanıtlarının ortak zarfı. Gerçek örnek değerler (message,
        statusCode, status) her status koduna özgü olarak ilgili
        components.responses girdisinde (ör. NotFound, Forbidden) tanımlıdır;
        bu şema yalnızca alan yapısını tarif eder.
      required: [success, message, statusCode, status]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
          description: 'Kullanıcıya gösterilebilecek, insan tarafından okunabilir hata mesajı'
        statusCode:
          type: integer
          description: 'HTTP durum koduyla birebir aynı değer'
        status:
          type: string
          description: "statusCode'a karşılık gelen sabit; frontend'in message string'ini parse etmeden dallanabilmesi içindir"
          enum:
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - UNPROCESSABLE_ENTITY
            - TOO_MANY_REQUESTS
            - INTERNAL_SERVER_ERROR
            - SERVICE_UNAVAILABLE
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field: { type: string, example: 'email' }
                  message: { type: string, example: 'Email is required' }

    # ---------- Success envelope ----------
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        message: { type: string }

    # ---------- Domain schemas ----------
    Role:
      type: string
      enum: [USER, ADMIN]
...
    User:
      type: object
      properties:
        id: { type: string, readOnly: true }
        email: { type: string, format: email }
        fullName: { type: string }
        phone: { type: string }
        role:
          allOf:
            - $ref: '#/components/schemas/Role'
          readOnly: true
        createdAt:
          {
            type: string,
            format: date-time,
            readOnly: true,
            example: '2026-07-14T10:35:15Z',
          }
...
    UserEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties: { data: { $ref: '#/components/schemas/User' } }
...